fix(lore-0145): address Copilot review — perf, overflow, tokio features
Three drive-by fixes from Copilot review on PR #111:

1. `status::partition_row` — iterate the clamped `[first..=last]` per
   partition (<=64k) and hit `completed.contains(seq)`, instead of
   scanning the full `completed` HashSet per partition. Avoids
   O(partitions × completed) behaviour that would matter at full
   Soroban-history scale (~14M ledgers × hundreds of partitions).

2. `run::execute::total_range` — widen to u64 before the subtraction
   so `end == u32::MAX && start == 0` doesn't wrap. Soroban ranges
   don't hit that in practice; fix is free.

3. `backfill-runner/Cargo.toml` — add `fs`, `process`, `time` to
   tokio features. Workspace build got away with it because other
   crates turn them on; `cargo build -p backfill-runner` solo was
   broken without them.

+++ crates/backfill-runner/src/run.rs
@@ -97,7 +97,10 @@ pub async fn execute(
     // Sticky dashboard. Visual bar covers the full range and is pre-
     // bumped by `completed.len()` (handled inside `Dashboard::new`);
     // `timing` is scoped only to the work this run actually has to do.
-    let total_range = (end - start + 1) as u64;
+    // Widen to u64 before the arithmetic — `(end - start + 1) as u32`
+    // wraps when `end == u32::MAX && start == 0`. Soroban ranges don't
+    // hit that in practice, but the fix is free. Copilot review on PR #111.
+    let total_range = u64::from(end) - u64::from(start) + 1;
     let already_done = completed.len() as u64;
     let dashboard = Arc::new(Dashboard::new(total_range, already_done, mp));
